% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.barplot.R
\name{effect}
\alias{effect}
\title{A function that illustrates bar plot of tw o variables}
\usage{
effect(data, x, y, color)
}
\arguments{
\item{data}{the data selected}

\item{x}{numeric. numeric vector}

\item{y}{numeric. numeric vector}

\item{color}{color of the barplot}
}
\value{
barplot a bar plot is returned with selected color
}
\description{
The function below will show us the bar plot of tw ovariables and how they depend on each other
}
\examples{
alcohol_consumption <- drugs \%>\%  effect_drugs(drugs$country, drugs$consumption_alcohol, color = "blue")
cit.acid_effect <- wine \%>\% effect(wine$quality, wine$citric.acid, color = "orange")
}
\author{
Islam and Leyla
}
